refacto contants
