--- src/Constants.ts
@@ -1,108 +1,0 @@
-// Canvas constants
-const REFRESH_INTERVAL = 30
-const CANVAS_WIDTH = 1200
-const CANVAS_HEIGHT = 720
-const CANVAS_SIDE_MARGIN = 20
-const HARDWARE_BUTTON_WIDTH = 100
-const HARDWARE_BUTTON_MARGIN = 50
-const UPGRADE_JAUGE_LENGTH = 200
-const UPGRADE_JAUGE_HEIGHT = 20
-const UPGRADE_BUTTON_WIDTH = 20
-const JAUGE_FILL_COLOR = "rgba(255, 255, 255, 0.6)"
-const BOX_BOTTOM_POSITIONS = [
-    {
-        x:
-            CANVAS_WIDTH / 2 -
-            (HARDWARE_BUTTON_WIDTH +
-                HARDWARE_BUTTON_WIDTH / 2 +
-                HARDWARE_BUTTON_MARGIN),
-        y: CANVAS_HEIGHT - HARDWARE_BUTTON_WIDTH - HARDWARE_BUTTON_MARGIN,
-    },
-    {
-        x: CANVAS_WIDTH / 2 - HARDWARE_BUTTON_WIDTH / 2,
-        y: CANVAS_HEIGHT - HARDWARE_BUTTON_WIDTH - HARDWARE_BUTTON_MARGIN,
-    },
-    {
-        x:
-            CANVAS_WIDTH / 2 +
-            HARDWARE_BUTTON_WIDTH / 2 +
-            HARDWARE_BUTTON_MARGIN,
-        y: CANVAS_HEIGHT - HARDWARE_BUTTON_WIDTH - HARDWARE_BUTTON_MARGIN,
-    },
-]
-
-// Game constants
-const NUMBER_OF_LIVES = 3
-const INVINCIBILITY_TIME = 3
-const XP_BAR_LENGTH = 300
-const XP_BAR_HEIGHT = 10
-const NEXT_LEVEL_XP = 30
-const LASER_SHOOTING_RATE = 2
-const LASER_RANGE = 500
-const LASER_SHOT_SPEED = 1000
-const LASER_SHOT_LENGTH = 10
-const NEW_GAME_DELAY = 2
-
-// Ship constants
-const SHIP_WIDTH = 10
-const SHIP_LENGTH = 20
-const SHIP_MAX_SPEED = 3
-const SHIP_ACCELERATION = 3
-const SHIELD_RELOAD_TIME = 5
-const BOMB_RATE = 6
-const BOMB_EXPLOSION_DELAY = 3
-const BOMB_SIZE = 10
-
-// Asteroids constants
-const ASTEROID_MAX_SPEED = 100
-const ASTEROID_SIZE = 20
-const GIGA_ASTEROID_SIZE = 200
-const ASTEROID_JAGGEDNESS = 0.3
-const ASTEROID_PARTICULE_LENGTH = 2
-
-const PARTICULE_MIN_SPEED = 50
-const PARTICULE_MAX_SPEED = 200
-const PARTICULE_MAX_RANGE = 100
-
-export {
-    REFRESH_INTERVAL,
-    CANVAS_WIDTH,
-    CANVAS_HEIGHT,
-    CANVAS_SIDE_MARGIN,
-    HARDWARE_BUTTON_WIDTH as MENU_UPGRADE_WIDTH,
-    HARDWARE_BUTTON_MARGIN as MENU_UPGRADE_MARGIN,
-    UPGRADE_JAUGE_LENGTH,
-    UPGRADE_JAUGE_HEIGHT,
-    UPGRADE_BUTTON_WIDTH,
-    JAUGE_FILL_COLOR as JAUGE_COLOR,
-    BOX_BOTTOM_POSITIONS,
-    NUMBER_OF_LIVES,
-    INVINCIBILITY_TIME,
-    XP_BAR_LENGTH,
-    XP_BAR_HEIGHT,
-    NEXT_LEVEL_XP,
-    LASER_SHOOTING_RATE,
-    LASER_RANGE,
-    LASER_SHOT_SPEED,
-    LASER_SHOT_LENGTH,
-    NEW_GAME_DELAY,
-    SHIP_WIDTH,
-    SHIP_LENGTH,
-    SHIP_MAX_SPEED,
-    SHIP_ACCELERATION,
-    SHIELD_RELOAD_TIME,
-    BOMB_RATE,
-    BOMB_EXPLOSION_DELAY,
-    BOMB_SIZE,
-}
-
-export {
-    ASTEROID_MAX_SPEED,
-    ASTEROID_SIZE,
-    GIGA_ASTEROID_SIZE,
-    ASTEROID_JAGGEDNESS,
-    ASTEROID_PARTICULE_LENGTH,
-    PARTICULE_MIN_SPEED,
-    PARTICULE_MAX_SPEED,
-    PARTICULE_MAX_RANGE,
-}
