+++ src/constants/game.ts
@@ -1,0 +1,27 @@
+const REFRESH_INTERVAL = 30
+const NEXT_LEVEL_XP = 30
+const NEW_GAME_DELAY = 2
+
+const ASTEROID_MAX_SPEED = 100
+const ASTEROID_SIZE = 20
+const GIGA_ASTEROID_SIZE = 200
+const ASTEROID_JAGGEDNESS = 0.3
+const ASTEROID_PARTICULE_LENGTH = 2
+
+const PARTICULE_MIN_SPEED = 50
+const PARTICULE_MAX_SPEED = 200
+const PARTICULE_MAX_RANGE = 100
+
+export {
+    REFRESH_INTERVAL,
+    NEXT_LEVEL_XP,
+    NEW_GAME_DELAY,
+    ASTEROID_MAX_SPEED,
+    ASTEROID_SIZE,
+    GIGA_ASTEROID_SIZE,
+    ASTEROID_JAGGEDNESS,
+    ASTEROID_PARTICULE_LENGTH,
+    PARTICULE_MIN_SPEED,
+    PARTICULE_MAX_SPEED,
+    PARTICULE_MAX_RANGE,
+}
