+++ src/constants/ship.ts
@@ -1,0 +1,31 @@
+const SHIP_WIDTH = 10
+const SHIP_LENGTH = 20
+const SHIP_MAX_SPEED = 3
+const SHIP_ACCELERATION = 3
+const NUMBER_OF_LIVES = 3
+const INVINCIBILITY_TIME = 3
+const LASER_SHOOTING_RATE = 2
+const LASER_RANGE = 500
+const LASER_SHOT_SPEED = 1000
+const LASER_SHOT_LENGTH = 10
+const SHIELD_RELOAD_TIME = 5
+const BOMB_RATE = 6
+const BOMB_EXPLOSION_DELAY = 3
+const BOMB_SIZE = 10
+
+export {
+    SHIP_WIDTH,
+    SHIP_LENGTH,
+    SHIP_MAX_SPEED,
+    SHIP_ACCELERATION,
+    NUMBER_OF_LIVES,
+    INVINCIBILITY_TIME,
+    LASER_SHOOTING_RATE,
+    LASER_RANGE,
+    LASER_SHOT_SPEED,
+    LASER_SHOT_LENGTH,
+    SHIELD_RELOAD_TIME,
+    BOMB_RATE,
+    BOMB_EXPLOSION_DELAY,
+    BOMB_SIZE,
+}
